Move files up out of graph
