+++ strawberryfields/apps/dense.py
