+++ strawberryfields/apps/g_sample.py
