+++ strawberryfields/apps/resize.py
